Rename a2ui_extension to a2ui_agent
This major refactoring renames the python agent library from `a2ui_extension` to `a2ui_agent` to better reflect its purpose as a core agent library rather than just an extension.

- Rename directory `a2a_agents/python/a2ui_extension` -> `a2a_agents/python/a2ui_agent`.
- Refactor project structure:
  - Move extension source files to `src/a2ui/extension/` to support proper namespacing.
  - Move extension tests to `tests/extension/`.
- Update pyproject.toml:
  - Rename project to `a2ui-agent`.
  - Configure `hatchling` build target.
- Update README.md with new directory paths and testing instructions.
- Fix imports in source and test files to match the new `a2ui.extension` namespace.
- Update all sample agents (`contact_lookup`, `contact_multiple_surfaces`, `orchestrator`, `restaurant_finder`, `rizzcharts`) to use the new `a2ui-agent` package name and path.
- Update python imports in all samples from `a2ui.a2ui_extension` to `a2ui.extension.a2ui_extension`.

Verification:
- Manually run the orchestration agent and client
- All workflow are working as expected.
