+++ a2a_agents/python/a2ui_agent/pyproject.toml
@@ -1,7 +1,7 @@
 [project]
-name = "a2ui"
+name = "a2ui-agent"
 version = "0.1.0"
-description = "A2UI Extension"
+description = "A2UI Agent"
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
@@ -14,6 +14,9 @@ dependencies = [
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/a2ui"]
 
 [[tool.uv.index]]
 url = "https://pypi.org/simple"
