+++ samples/agent/adk/contact_multiple_surfaces/__main__.py
@@ -20,7 +20,7 @@ from a2a.server.apps import A2AStarletteApplication
 from a2a.server.request_handlers import DefaultRequestHandler
 from a2a.server.tasks import InMemoryTaskStore
 from a2a.types import AgentCapabilities, AgentCard, AgentSkill
-from a2ui.a2ui_extension import get_a2ui_agent_extension
+from a2ui.extension.a2ui_extension import get_a2ui_agent_extension
 from agent import ContactAgent
 from agent_executor import ContactAgentExecutor
 from dotenv import load_dotenv
