+++ samples/agent/adk/pyproject.toml
@@ -6,5 +6,5 @@ default = true
 members = ["contact_lookup", "contact_multiple_surfaces", "orchestrator", "restaurant_finder", "rizzcharts"]
 
 [tool.uv.sources]
-a2ui = { path = "../../../a2a_agents/python/a2ui_extension", editable = true }
+a2ui-agent = { path = "../../../a2a_agents/python/a2ui_agent", editable = true }
