+++ samples/agent/adk/restaurant_finder/agent_executor.py
@@ -32,7 +32,7 @@ from a2a.utils import (
     new_task,
 )
 from a2a.utils.errors import ServerError
-from a2ui.a2ui_extension import create_a2ui_part, try_activate_a2ui_extension
+from a2ui.extension.a2ui_extension import create_a2ui_part, try_activate_a2ui_extension
 from agent import RestaurantAgent
 
 logger = logging.getLogger(__name__)
