+++ samples/agent/adk/restaurant_finder/pyproject.toml
@@ -12,7 +12,7 @@ dependencies = [
     "python-dotenv>=1.1.0",
     "litellm",
     "jsonschema>=4.0.0",
-    "a2ui",
+    "a2ui-agent",
 ]
 
 [tool.hatch.build.targets.wheel]
